fix again (smh)

+++ commands/top.js
@@ -63,7 +63,7 @@ module.exports = {
             networth.sort((a, b) => b.networth - a.networth)
 				.filter(user => interaction.guild.members.cache.has(user.user_id))
                 .first(30)
-                .map((user, position) => `(${position + 1}) ${(interaction.guild.members.cache.get(user.user_id).user.username)}: ${parseInt(user.networth).toFixed(2)} ⵇ`)
+                .map((user, position) => `(${position + 1}) ${(interaction.guild.members.cache.get(user.user_id).user.username)}: ${parseFloat(user.networth).toFixed(2)} ⵇ`)
                 .join('\n'),
             { code: true }
         );
